+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 - **Changes**: --inverse option to show common commits [#118][] [#134][]
 - **Snapshot**: replace a snapshot with --replace [#143][]
 - **State**: add --show-all option [#145][]
+- **Settings**: --format now supports count, keys, and sections
 
 ### Changes
 - **Settings**: remove get command [#135][]